+++ erkenntnis/brain_implementation/preprocess_perception.py
@@ -21,7 +21,7 @@ def split_perception_by_type(perception):
 
 def fuse_state(encoded_perception, encoded_messages, encoded_action, encoded_cause):
     if encoded_cause is None:
-        encoded_cause = np.array([0.0])
+        encoded_cause = np.array([0.0], dtype=np.float32)
     fused_state = np.concatenate([encoded_perception, encoded_messages, encoded_action, encoded_cause])
     # fused_state = np.reshape(fused_state, (1, fused_state.shape[0]))
     return fused_state
